[assistant] - refactor: streamline AssistantDetails component structure
 - Extracted performance related code from AssistantDetails.tsx into a new AssistantDetailsPerformance.tsx component
 - Removed dropdown related code and period constants from AssistantDetails.tsx as they are now part of AssistantDetailsPerformance
 - Updated import statements in various files to accommodate the restructuring of AssistantDetails and AssistantDetailsPerformance components
 - Adjusted AssistantBuilder.tsx to pass the entire agentConfiguration object when invoking other components instead of just the agentConfigurationId
 - Modified AssistantBuilderPreviewDrawer.tsx to use the new AssistantDetailsPerformance component and pass the required gridMode prop
 - Removed the direct FeedbacksSection import and usage from the AssistantBuilderPreviewDrawer since it's now part of AssistantDetailsPerformance
 - Updated types and props in several components to align with changes made in the AssistantDetails and AssistantDetailsPerformance components
