+++ front/components/assistant_builder/AssistantBuilderPreviewDrawer.tsx
@@ -24,7 +24,6 @@ import { ConversationsNavigationProvider } from "@app/components/assistant/conve
 import ConversationViewer from "@app/components/assistant/conversation/ConversationViewer";
 import { GenerationContextProvider } from "@app/components/assistant/conversation/GenerationContextProvider";
 import { AssistantInputBar } from "@app/components/assistant/conversation/input_bar/InputBar";
-import { FeedbacksSection } from "@app/components/assistant_builder/FeedbacksSection";
 import {
   usePreviewAssistant,
   useTryAssistantCore,
@@ -46,6 +45,8 @@ import type {
   WorkspaceType,
 } from "@app/types";
 import { isAssistantBuilderRightPanelTab } from "@app/types";
+import type { AgentConfigurationType } from "@app/types";
+import { AssistantDetailsPerformance } from "@app/components/assistant/AssistantDetailsPerformance";
 
 interface AssistantBuilderRightPanelProps {
   screen: BuilderScreen;
@@ -55,7 +56,7 @@ interface AssistantBuilderRightPanelProps {
   resetToTemplateActions: () => Promise<void>;
   owner: WorkspaceType;
   builderState: AssistantBuilderState;
-  agentConfigurationId: string | null;
+  agentConfiguration: AgentConfigurationType | null;
   setAction: (action: AssistantBuilderSetActionType) => void;
   reasoningModels: ModelConfigurationType[];
 }
@@ -68,7 +69,7 @@ export default function AssistantBuilderRightPanel({
   resetToTemplateActions,
   owner,
   builderState,
-  agentConfigurationId,
+  agentConfiguration,
   setAction,
   reasoningModels,
 }: AssistantBuilderRightPanelProps) {
@@ -129,7 +130,7 @@ export default function AssistantBuilderRightPanel({
               icon={ChatBubbleBottomCenterTextIcon}
             />
             {/* The agentConfigurationId is truthy if not a new agent */}
-            {agentConfigurationId && (
+            {agentConfiguration && (
               <TabsTrigger
                 value="Performance"
                 label="Performance"
@@ -264,12 +265,11 @@ export default function AssistantBuilderRightPanel({
             </div>
           </div>
         )}
-        {rightPanelTab === "Performance" && agentConfigurationId && (
-          <div className="ml-4 mt-4">
-            <Page.SectionHeader title="Feedback" />
-            <FeedbacksSection
+        {rightPanelTab === "Performance" && agentConfiguration && (
+          <div className="flex flex-col gap-5 pt-6 text-sm text-foreground dark:text-foreground-night">
+            <AssistantDetailsPerformance
+              agentConfiguration={agentConfiguration}
               owner={owner}
-              agentConfigurationId={agentConfigurationId}
               gridMode
             />
           </div>
